tributraries should happen when using the river brush
